+++ core/network/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+    alias(libs.plugins.koin.library)
+    alias(libs.plugins.koin.hilt)
+}
+
+android {
+    namespace = "in.koreatech.koin.core.network"
+}
+
+dependencies {
+    implementation(libs.androidx.core.ktx)
+    implementation(libs.androidx.appcompat)
+}
