+++ core/network/src/main/java/com/example/network/service/NetworkConnectivityService.kt
@@ -1,12 +1,11 @@
-package `in`.koreatech.koin.data.service
+package com.example.network.service
 
 import android.net.ConnectivityManager
 import android.net.ConnectivityManager.NetworkCallback
 import android.net.Network
 import android.net.NetworkCapabilities
 import android.net.NetworkRequest
-import `in`.koreatech.koin.domain.service.NetworkConnectivityService
-import `in`.koreatech.koin.domain.state.network.NetworkStatus
+import com.example.network.state.NetworkStatus
 import javax.inject.Inject
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.channels.awaitClose
@@ -14,6 +13,11 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.callbackFlow
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.flowOn
+
+interface NetworkConnectivityService {
+    val networkStatus: Flow<NetworkStatus>
+    fun getLatestStatus(): NetworkStatus
+}
 
 class NetworkConnectivityServiceImpl @Inject constructor(
     connectivityManager: ConnectivityManager
