+++ core/network/src/test/java/com/example/network/ExampleUnitTest.kt
@@ -1,0 +1,16 @@
+package `in`.koreatech.koin.core.navigation
+
+import org.junit.Assert.assertEquals
+import org.junit.Test
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+class ExampleUnitTest {
+    @Test
+    fun addition_isCorrect() {
+        assertEquals(4, 2 + 2)
+    }
+}
