+++ koin/src/main/java/in/koreatech/koin/KoinApplication.kt
@@ -10,7 +10,6 @@ import `in`.koreatech.koin.core.toast.ToastUtil
 import `in`.koreatech.koin.data.sharedpreference.RecentSearchSharedPreference
 import `in`.koreatech.koin.data.sharedpreference.UserInfoSharedPreferencesHelper
 import `in`.koreatech.koin.domain.repository.TokenRepository
-import `in`.koreatech.koin.domain.service.NetworkConnectivityService
 import `in`.koreatech.koin.domain.usecase.user.GetLoggerUserDataUseCase
 import `in`.koreatech.koin.util.ExceptionHandlerUtil
 import javax.inject.Inject
@@ -27,9 +26,6 @@ class KoinApplication : Application() {
     @Inject
     lateinit var getLoggerUserDataUseCase: GetLoggerUserDataUseCase
 
-    @Inject
-    lateinit var networkConnectivityService: NetworkConnectivityService
-
     override fun onCreate() {
         super.onCreate()
         init()
@@ -39,13 +35,7 @@ class KoinApplication : Application() {
         UserInfoSharedPreferencesHelper.getInstance().init(applicationContext)
         ToastUtil.getInstance().init(applicationContext)
         RecentSearchSharedPreference.getInstance().init(applicationContext)
-        Thread.setDefaultUncaughtExceptionHandler(
-            ExceptionHandlerUtil(
-                applicationContext,
-                networkConnectivityService.networkStatus,
-                networkConnectivityService.getLatestStatus()
-            )
-        )
+        Thread.setDefaultUncaughtExceptionHandler(ExceptionHandlerUtil(applicationContext))
         initTimber()
         KakaoSdk.init(this, BuildConfig.KAKAO_NATIVE_APP_KEY)
         EventLogger.init(getLoggerUserDataUseCase)
