+++ koin/src/main/java/in/koreatech/koin/util/ExceptionHandlerUtil.kt
@@ -6,7 +6,6 @@ import android.os.Looper
 import androidx.core.os.HandlerCompat
 import com.google.firebase.crashlytics.FirebaseCrashlytics
 import `in`.koreatech.koin.R
-import `in`.koreatech.koin.domain.state.network.NetworkStatus
 import `in`.koreatech.koin.feature.user.ui.signin.SignInActivity
 import `in`.koreatech.koin.ui.error.ErrorActivity
 import `in`.koreatech.koin.util.ext.showToast
@@ -14,27 +13,8 @@ import java.io.PrintWriter
 import java.io.StringWriter
 import java.io.Writer
 import kotlin.system.exitProcess
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.SupervisorJob
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.SharingStarted.Companion.WhileSubscribed
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.stateIn
 
-class ExceptionHandlerUtil(
-    private val context: Context,
-    private val networkStatusFlow: Flow<NetworkStatus>,
-    private val initNetworkStatus: NetworkStatus
-) : Thread.UncaughtExceptionHandler {
-
-    private val scope = CoroutineScope(SupervisorJob() + Dispatchers.Default)
-
-    private val networkStatus: StateFlow<NetworkStatus> = networkStatusFlow.stateIn(
-        initialValue = initNetworkStatus,
-        scope = scope,
-        started = WhileSubscribed(5000)
-    )
+class ExceptionHandlerUtil(private val context: Context) : Thread.UncaughtExceptionHandler {
 
     /***
      * UncaughtException을 캐치하여 처리하는 함수
@@ -43,12 +23,6 @@ class ExceptionHandlerUtil(
      * @param throwable
      */
     override fun uncaughtException(thread: Thread, throwable: Throwable) {
-        if (networkStatus.value == NetworkStatus.Disconnected) {
-            context.applicationContext.showToast(
-                context.getString(R.string.error_network_connection)
-            )
-            return
-        }
         FirebaseCrashlytics.getInstance().recordException(throwable)
 
         val stringWriter = StringWriter()
