+++ settings.gradle
@@ -34,3 +34,4 @@ include ':feature:chat'
 include ':feature:banner'
 include ':feature:user'
 include ':feature:club'
+include ':core:network'
